// src/types/Prediction.ts
export declare namespace Prediction {
  export type BetInfoStruct = {
    position: BigNumberish;
    token: string;
    amount: BigNumberish;
    claimed: boolean;
  };

  export type BetInfoStructOutput = [number, string, BigNumber, boolean] & {
    position: number;
    token: string;
    amount: BigNumber;
    claimed: boolean;
  };
}